+++ src/fable-library-ts/Native.fs
@@ -94,3 +94,5 @@ module Helpers =
     // Using Emit keeps the argument signature
     [<Emit("$1.sort($0)")>]
     let sortInPlaceWithImpl (comparer: 'T -> 'T -> int) (array: 'T[]) : unit = nativeOnly
+
+    let inline arrayAccessImpl (index: int) (array: 'T[]) : 'T = emitJsExpr (index, array) "$1[$0]"